dotnet: Fix BlurgCursor def

+++ dotnet/BlurgText/BlurgCursor.cs
@@ -7,5 +7,5 @@ public struct BlurgCursor
 {
     public int X;
     public int Y;
-    public float Height;
+    public int Height;
 }
